% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{age_groups}
\alias{age_groups}
\title{Female Age Group Categories}
\format{
A vector with 8 observations.
\describe{
\item{age_groups}{Age Groups For Female Earnings Dataset}
}
}
\source{
\url{https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2019/2019-03-05/earnings_female.csv}
}
\usage{
age_groups
}
\description{
Female Age Group Categories
}
\keyword{datasets}
