% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earnings_female}
\alias{earnings_female}
\title{Women's Earnings Throughout The Years}
\format{
A data frame with 264 observations and 3 variables.
\describe{
\item{Year}{Female Earnings For A Particular Year}
\item{Age_Group}{Various Age Groups of Females}
\item{percent}{Percentage Earnings of Females}
}
}
\source{
\url{https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2019/2019-03-05/earnings_female.csv}
}
\usage{
earnings_female
}
\description{
Women's Earnings Throughout The Years
}
\keyword{datasets}
