% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/earningsplot.R
\name{earningsplot}
\alias{earningsplot}
\title{Line Plot For Female Earnings For Particular Age Group}
\usage{
earningsplot(age_group)
}
\arguments{
\item{age_group}{A string containing age group of females}
}
\value{
A line plot of entered age group
}
\description{
Generates a line plot for percentage earnings made by females throughout
1979-2011 for the particular age group entered.
}
\examples{
earningsplot("16-19 years")

}
