% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{employed_gender}
\alias{employed_gender}
\title{Employment Rate For Both Males and Females}
\format{
A data frame with 196 observations on 3 variables .
\describe{
\item{year}{Employment Rate For A Particular Year}
\item{Job_Type}{Types of Job - Part-Time/Full-Time}
\item{Percentage}{Percentage Employment Rate For Each Gender}
}
}
\source{
\url{https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2019/2019-03-05/employed_gender.csv}
}
\usage{
employed_gender
}
\description{
Employment Rate For Both Males and Females
}
\keyword{datasets}
