% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jobs_gender}
\alias{jobs_gender}
\title{Workforce, Earnings For Each Gender By Occupation}
\format{
A data frame with 2088 observations on 12 variables .
\describe{
\item{year}{Earnings For A Job Category For That Particular Year}
\item{occupation}{Specific Job Name}
\item{major_category}{Top-Level/Broad Category Of Occupation}
\item{minor_category}{Lower-Level/Fine Category Of Occupation}
\item{total_workers}{Total Estimated Full-Time Workers}
\item{workers_male}{Estimated Male Full-Time Workers}
\item{workers_female}{Estimated Female Full-Time Workers}
\item{percent_female}{Percentage Of Females For Specific Occupation}
\item{total_earnings}{Total Estimated Earnings For Full-Time Workers}
\item{total_earnings_male}{Total Estimated Earnings For Male Full-Time Workers}
\item{total_earnings_female}{Total Estimated Earnings For Female Full-Time Workers}
\item{wage_percent_of_male}{Female Wages As Percent Of Male Wages}
}
}
\source{
\url{https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2019/2019-03-05/jobs_gender.csv}
}
\usage{
jobs_gender
}
\description{
Workforce, Earnings For Each Gender By Occupation
}
\keyword{datasets}
