% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{jobs_workforce}
\alias{jobs_workforce}
\title{Workforce For Each Gender By Occupation}
\format{
A data frame with 4176 observations on 5 variables .
\describe{
\item{year}{Earnings For A Job Category For That Particular Year}
\item{occupation}{Specific Job Name}
\item{minor_category}{Lower-Level/Fine Category Of Occupation}
\item{Gender}{Male or Female}
\item{Workforce}{Total Workforce Present}
}
}
\source{
\url{https://raw.githubusercontent.com/rfordatascience/tidytuesday/master/data/2019/2019-03-05/jobs_gender.csv}
}
\usage{
jobs_workforce
}
\description{
Workforce For Each Gender By Occupation
}
\keyword{datasets}
