% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quadratic.R
\name{quadratic}
\alias{quadratic}
\title{Roots of A Quadratic Equation}
\usage{
quadratic(a, b, c)
}
\arguments{
\item{a}{A real numeric coefficient}

\item{b}{A real numeric coefficient}

\item{c}{A real numeric coefficient}
}
\value{
A tibble consisting of 2 variables, \code{Roots} and \code{Root_Type}
}
\description{
Calculates the roots of a quadratic equation and returns either 2 distinct
real roots, 1 repeated real root, or 2 imaginary roots depending on the
discriminant calculated. The order of the roots entered is important as each
input variable corresponds to the coefficient in a simple quadratic equation.
The values \verb{a,b,c} are coefficients of the equation and follow the
nomenclature of \code{ax^2 + bx + c} where \code{a!=0}.
}
\examples{
quadratic(1,5,6)
quadratic(1,-6,9)
quadratic(5,2,1)

}
