% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run-app.R
\name{run_app}
\alias{run_app}
\title{Shiny App}
\usage{
run_app()
}
\value{
Shiny App
}
\description{
Launches the Shiny App For the Project, Women In The Workplace where the
user is able to gain insights on how women perform in the workplace with
respect to earnings, employment rate, and when various job categories are
factored in. It also shows how they stack up against men for comparison
and whether gender discrimination is prevalent or not. The app also allows
the user to interact with the visualizations and filter the plots depending
on the inputs selected.
}
