% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WomInWorkShiny-package.R
\docType{package}
\name{WomInWorkShiny-package}
\alias{WomInWorkShiny}
\alias{WomInWorkShiny-package}
\title{WomInWorkShiny: Runs An Interactive Shiny App Along With A Quadratic Function}
\description{
Allows the user to explore various datasets related to women's earnings, overall employment rate, and the employment rate by job category. It also launches an inbuilt Shiny app to visualize and interact with the plots to gain valuable insights. It also has a function to find the roots of a quadratic equation depending on the discriminant calculated.
}
\author{
\strong{Maintainer}: Shreyansh Mahtolia \email{smah0071@student.monash.edu}

}
\keyword{internal}
